[NFC] Move PinsStore to PackageGraph
This will allow us to access PinsStore inside the dependency resolver.
