+++ Sources/PackageGraph/CheckoutState.swift
@@ -9,7 +9,6 @@
  */
 
 import Basic
-import PackageGraph
 import SourceControl
 import SPMUtility
 
@@ -52,7 +51,7 @@ public struct CheckoutState: Equatable, CustomStringConvertible {
         return version?.description ?? branch ?? revision.identifier
     }
 
-    var isBranchOrRevisionBased: Bool {
+    public var isBranchOrRevisionBased: Bool {
         return version == nil
     }
 }
@@ -60,7 +59,7 @@ public struct CheckoutState: Equatable, CustomStringConvertible {
 extension CheckoutState {
 
     /// Returns requirement induced by this state.
-    func requirement() -> PackageRequirement {
+    public func requirement() -> PackageRequirement {
         if let version = version {
             return .versionSet(.exact(version))
         } else if let branch = branch {
@@ -87,15 +86,6 @@ extension CheckoutState: JSONMappable, JSONSerializable {
            "version": version.toJSON(),
            "branch": branch.toJSON(),
        ])
-    }
-}
-
-extension ManagedDependency {
-    public var checkoutState: CheckoutState? {
-        if case .checkout(let checkoutState) = state {
-            return checkoutState
-        }
-        return nil
     }
 }
